Add files via upload
Oh lets go question text pops up now and I got rid of the extra spacing issue. Actually looks semi functional now. Granted its not. From here we need to also send the text content and all the other whatever info to the database so that the TA view is able to view it concurrently on their view. Creating the TA view of this should be relatively simple since its just a different version of the student page.